added covert method to fda/fir.py and deleted excess files
